requirements.txt
absl-py==0.9.0
aiofiles==0.5.0
aiohttp==3.6.2
APScheduler==3.6.3
astor==0.8.1
async-generator==1.10
async-timeout==3.0.1
attrs==19.3.0
boto3==1.13.16
botocore==1.16.16
bz2file==0.98
cachetools==4.1.0
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.2
cloudpickle==1.2.2
colorclass==2.2.0
coloredlogs==10.0
colorhash==1.0.2
conda==4.8.3
conda-package-handling==1.7.0
ConfigArgParse==1.2.3
cryptography==2.9.2
cycler==0.10.0
decorator==4.4.2
dill==0.3.1.1
dnspython==1.16.0
docopt==0.6.2
docutils==0.15.2
dopamine-rl==3.0.1
fbmessenger==6.0.0
Flask==1.1.2
future==0.18.2
gast==0.2.2
gevent==1.4.0
gin-config==0.3.0
google-api-core==1.17.0
google-api-python-client==1.8.3
google-auth==1.15.0
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
googleapis-common-protos==1.51.0
greenlet==0.4.15
grpcio==1.29.0
gunicorn==20.0.4
gym==0.15.4
h11==0.8.1
h2==3.2.0
h5py==2.10.0
hpack==3.0.0
hstspreload==2020.5.19
httpcore==0.3.0
httplib2==0.18.1
httptools==0.1.1
httpx==0.9.3
humanfriendly==8.2
hyperframe==5.2.0
idna==2.9
importlib-metadata==1.6.0
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.10.0
joblib==0.15.1
jsonpickle==1.3
jsonschema==3.2.0
kafka-python==1.4.7
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kfac==0.2.0
kiwisolver==1.2.0
Markdown==3.2.2
MarkupSafe==1.1.1
matplotlib==3.1.3
mattermostwrapper==2.2
mesh-tensorflow==0.1.13
mpmath==1.1.0
multidict==4.6.1
networkx==2.4
nltk==3.5
numpy==1.18.4
oauth2client==4.1.3
oauthlib==3.1.0
opencv-python==4.2.0.34
opt-einsum==3.2.1
packaging==19.0
pandas==1.0.3
pika==1.0.1
Pillow==7.1.2
promise==2.3
prompt-toolkit==2.0.10
protobuf==3.12.1
psycopg2-binary==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycosat==0.6.3
pycparser==2.20
pydot==1.4.1
pyglet==1.3.2
PyJWT==1.7.1
pykwalify==1.7.0
pymongo==3.8.0
pyOpenSSL==19.1.0
pyparsing==2.4.7
pypng==0.0.20
pyrsistent==0.16.0
PySocks==1.7.1
python-crfsuite==0.9.7
python-dateutil==2.8.1
python-dotenv==0.13.0
python-engineio==3.11.2
python-socketio==4.4.0
python-telegram-bot==12.7
pytz==2019.3
PyYAML==5.3.1
questionary==1.5.2
rasa==1.7.3
rasa-sdk==1.7.0
redis==3.3.11
regex==2020.5.14
requests==2.23.0
requests-async==0.5.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
rfc3986==1.4.0
rocketchat-API==0.6.36
rsa==4.0
ruamel-yaml==0.15.87
s3transfer==0.3.3
sanic==19.9.0
Sanic-Cors==0.9.9.post1
sanic-jwt==1.3.2
Sanic-Plugins-Framework==0.9.2
scikit-learn==0.20.4
scipy==1.4.1
six==1.14.0
sklearn-crfsuite==0.3.6
slackclient==1.3.2
sniffio==1.1.0
SQLAlchemy==1.3.17
sympy==1.5.1
tabulate==0.8.7
tensor2tensor==1.14.1
tensorboard==1.15.0
tensorflow==2.1.1
tensorflow-addons==0.8.3
tensorflow-cpu==1.15.0
tensorflow-datasets==3.1.0
tensorflow-estimator==1.15.1
tensorflow-gan==2.0.0
tensorflow-hub==0.7.0
tensorflow-metadata==0.22.0
tensorflow-probability==0.7.0
termcolor==1.1.0
terminaltables==3.1.0
tornado==6.0.4
tqdm==4.31.1
twilio==6.26.3
typeguard==2.7.1
tzlocal==2.1
ujson==1.35
uritemplate==3.0.1
urllib3==1.25.8
uvloop==0.14.0
wcwidth==0.1.9
webexteamssdk==1.1.1
websocket-client==0.54.0
websockets==8.1
Werkzeug==1.0.1
wrapt==1.12.1
yarl==1.4.2
zipp==3.1.0
